Moves AttributeDict into new util module.
I've used the AttributeDict wrapper in a couple other projects,
and each time it's bothered me that it's under clik.app.
`from clik.util import AttributeDict` seems a lot more right than
`from clik.app import AttributeDict`. Hence this change.
