+++ src/clik/app.py
@@ -12,6 +12,7 @@ from clik.argparse import ALLOW_UNKNOWN, ArgumentParser, ArgumentParserExit
 from clik.command import Command
 from clik.context import Context
 from clik.magic import Magic
+from clik.util import AttributeDict
 
 
 #: Magic variable containing parsed arguments.
@@ -76,38 +77,6 @@ def app(fn=None, name=None):
     if fn is None:
         return decorate
     return decorate(fn)
-
-
-class AttributeDict(dict):
-    """
-    Simple :class:`dict` wrapper that allows key access via attribute.
-
-    Example::
-
-        d = AttributeDict(foo='bar', baz='qux')
-        d['foo']      # 'bar'
-        d.foo         # 'bar'
-        d['baz']      # 'qux'
-        d.baz         # 'qux'
-        d.foo = 'bup'
-        d['foo']      # 'bup'
-        d.foo         # 'bup'
-        del d.foo
-        d.foo         # KeyError
-
-    """
-
-    def __getattr__(self, name):
-        """Get via attribute name."""
-        return self[name]
-
-    def __setattr__(self, name, value):
-        """Set via attribute name."""
-        self[name] = value
-
-    def __delattr__(self, name):
-        """Delete via attribute name."""
-        del self[name]
 
 
 class App(Command):
